reorganised components and moved calendar into separate folder
